langchain/deepseekchat/requirements.txt
langchain==0.3.19
langchain_deepseek==0.1.2
python-dotenv==1.0.1
